fix sca problems for metadata

+++ pkg/vm/engine/aoe/storage/logstore/store_test.go
@@ -73,9 +73,9 @@ func mockETDDLHandler(r io.Reader, meta *EntryMeta) (Entry, int64, error) {
 	entry := newEmptyDDLEntry(meta)
 	n, err := entry.ReadFrom(r)
 	if err != nil {
-		return nil, int64(n), err
+		return nil, n, err
 	}
-	return entry, int64(n), err
+	return entry, n, err
 }
 
 func TestStore(t *testing.T) {
